Initial import of debian/ folder from SVN repo
We are migrating the 'debian packaging' from the debian SVN repo to
debian GIT repo. From now on, the GST debian packages will be built
using git-buildpackage methodology (e.g. upstream, master, ubuntu
branches...)

It was agreed not to import the complete history of GST debian packages
and instead make a fresh start.

Signed-off-by: Nicolas Dechesne <n-dechesne@ti.com>